Added my old Java work
According to dropbox most of it is from 2012. The package declaratios
are wrong for the folder that they are currently located in, but that
is something that can be fixed at a later time when the work in java
is resumed (perhaps at the next Project Euler night with PVV?).